refactor: internal imports inside `nuxt` to use `nuxt/schema`
