+++ packages/nuxt/config.d.ts
@@ -1,4 +1,4 @@
-import type { NuxtConfig } from '@nuxt/schema'
-export { NuxtConfig } from '@nuxt/schema'
+import type { NuxtConfig } from 'nuxt/schema'
+export { NuxtConfig } from 'nuxt/schema'
 
 export declare function defineNuxtConfig(config: NuxtConfig): NuxtConfig
