+++ packages/nuxt/src/app/components/nuxt-island.ts
@@ -1,10 +1,10 @@
 import { defineComponent, createStaticVNode, computed, ref, watch } from 'vue'
 import { debounce } from 'perfect-debounce'
 import { hash } from 'ohash'
-import type { MetaObject } from '@nuxt/schema'
 import { appendHeader } from 'h3'
 // eslint-disable-next-line import/no-restricted-paths
 import type { NuxtIslandResponse } from '../../core/runtime/nitro/renderer'
+import type { MetaObject } from 'nuxt/schema'
 import { useNuxtApp } from '#app/nuxt'
 import { useRequestEvent } from '#app/composables/ssr'
 import { useHead } from '#app/composables/head'
