+++ packages/nuxt/src/components/scan.ts
@@ -1,11 +1,11 @@
 import { basename, extname, join, dirname, relative } from 'pathe'
 import { globby } from 'globby'
 import { pascalCase, splitByCase } from 'scule'
-import type { Component, ComponentsDir } from '@nuxt/schema'
 import { isIgnored } from '@nuxt/kit'
 // eslint-disable-next-line vue/prefer-import-from-vue
 import { hyphenate } from '@vue/shared'
 import { withTrailingSlash } from 'ufo'
+import type { Component, ComponentsDir } from 'nuxt/schema'
 
 /**
  * Scan the components inside different components folders
