+++ packages/nuxt/src/components/templates.ts
@@ -1,6 +1,6 @@
 import { isAbsolute, relative } from 'pathe'
-import type { Component, Nuxt, NuxtPluginTemplate, NuxtTemplate } from '@nuxt/schema'
 import { genDynamicImport, genExport, genImport, genObjectFromRawEntries } from 'knitwork'
+import type { Component, Nuxt, NuxtPluginTemplate, NuxtTemplate } from 'nuxt/schema'
 
 export interface ComponentsTemplateContext {
   nuxt: Nuxt
