+++ packages/nuxt/src/core/builder.ts
@@ -1,9 +1,9 @@
 import chokidar from 'chokidar'
-import type { Nuxt } from '@nuxt/schema'
 import { importModule, isIgnored } from '@nuxt/kit'
 import { debounce } from 'perfect-debounce'
 import { normalize } from 'pathe'
 import { createApp, generateApp as _generateApp } from './app'
+import type { Nuxt } from 'nuxt/schema'
 
 export async function build (nuxt: Nuxt) {
   const app = createApp(nuxt)
