+++ packages/nuxt/src/core/plugins/import-protection.ts
@@ -2,8 +2,8 @@ import { createRequire } from 'node:module'
 import { createUnplugin } from 'unplugin'
 import { logger } from '@nuxt/kit'
 import { isAbsolute, join, relative } from 'pathe'
-import type { Nuxt } from '@nuxt/schema'
 import escapeRE from 'escape-string-regexp'
+import type { Nuxt } from 'nuxt/schema'
 
 const _require = createRequire(import.meta.url)
