+++ packages/nuxt/src/head/runtime/index.ts
@@ -1,5 +1,5 @@
 import type { UseHeadInput } from '@vueuse/head'
-import type { HeadAugmentations } from '@nuxt/schema'
+import type { HeadAugmentations } from 'nuxt/schema'
 
 export * from './composables'
