+++ packages/nuxt/src/imports/module.ts
@@ -2,9 +2,9 @@ import { addVitePlugin, addWebpackPlugin, defineNuxtModule, addTemplate, resolve
 import { isAbsolute, join, relative, resolve, normalize } from 'pathe'
 import type { Import, Unimport } from 'unimport'
 import { createUnimport, scanDirExports } from 'unimport'
-import type { ImportsOptions, ImportPresetWithDeprecation } from '@nuxt/schema'
 import { TransformPlugin } from './transform'
 import { defaultPresets } from './presets'
+import type { ImportsOptions, ImportPresetWithDeprecation } from 'nuxt/schema'
 
 export default defineNuxtModule<Partial<ImportsOptions>>({
   meta: {
