+++ packages/nuxt/test/scan-components.test.ts
@@ -1,7 +1,7 @@
 import { resolve } from 'node:path'
-import type { ComponentsDir } from '@nuxt/schema'
 import { expect, it, vi } from 'vitest'
 import { scanComponents } from '../src/components/scan'
+import type { ComponentsDir } from 'nuxt/schema'
 
 const fixtureDir = resolve(__dirname, 'fixture')
 const rFixture = (...p: string[]) => resolve(fixtureDir, ...p)
