+++ packages/nuxt/types.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="nitropack" />
 export * from './dist/index'
 
-import type { SchemaDefinition } from '@nuxt/schema'
+import type { SchemaDefinition } from 'nuxt/schema'
 
 declare global {
   const defineNuxtConfig: typeof import('nuxt/config')['defineNuxtConfig']
